update logs storage architecture
