+++ gateway/core/services/storage/logs_storage.py
@@ -1,132 +1,24 @@
-"""
-This module handles the access to the logs store.
-"""
+"""Abstract base class for logs storage."""
 
-import logging
-import os
+from abc import ABC, abstractmethod
 from typing import Optional
 
-from core.models import Job
-from core.services.storage.enums.working_dir import WorkingDir
-from core.services.storage.path_builder import PathBuilder
 
-logger = logging.getLogger("core.LogsStorage")
+class LogsStorage(ABC):
+    """Abstract interface for job logs storage."""
 
-
-class LogsStorage:
-    """
-    Manages access to logs generated by a Job.
-
-    Logs are stored in two locations:
-    - Public logs belong to the user (USER_STORAGE)
-    - Private logs belong to the provider (PROVIDER_STORAGE)
-    """
-
-    FILE_EXTENSION = ".log"
-    PATH = "logs"
-    ENCODING = "utf-8"
-
-    def __init__(self, job: Job) -> None:
-        self._job_id = str(job.id)
-        self._user_id = job.author_id
-        self._username = job.author.username
-        self._function_title = job.program.title
-        self._provider_name = job.program.provider.name if job.program.provider else None
-
-        # Build public logs path (USER_STORAGE)
-        # User job: {author}/logs/
-        # Provider job: {author}/{provider}/{function}/logs/
-        self._public_path = PathBuilder.absolute_path(
-            working_dir=WorkingDir.USER_STORAGE,
-            username=self._username,
-            function_title=self._function_title,
-            provider_name=self._provider_name,
-            extra_sub_path=self.PATH,
-        )
-
-        # Build private logs path (PROVIDER_STORAGE)
-        # Path: {provider}/{function}/logs/
-        if self._provider_name:
-            self._private_path = PathBuilder.absolute_path(
-                working_dir=WorkingDir.PROVIDER_STORAGE,
-                username=self._username,
-                function_title=self._function_title,
-                provider_name=self._provider_name,
-                extra_sub_path=self.PATH,
-            )
-        else:
-            self._private_path = None
-
+    @abstractmethod
     def get_public_logs(self) -> Optional[str]:
         """Retrieve public logs for the job."""
-        return self._read_logs(self._public_path)
 
+    @abstractmethod
     def save_public_logs(self, logs: str) -> None:
-        """Save public logs for the job."""
-        self._write_logs(self._public_path, logs)
+        """Persist public logs for the job."""
 
+    @abstractmethod
     def get_private_logs(self) -> Optional[str]:
         """Retrieve private logs for the job (provider jobs only)."""
-        if self._private_path is None:
-            raise RuntimeError("Private logs are only available for provider jobs")
-        return self._read_logs(self._private_path)
 
+    @abstractmethod
     def save_private_logs(self, logs: str) -> None:
-        """Save private logs for the job (provider jobs only)."""
-        if self._private_path is None:
-            raise RuntimeError("Private logs are only available for provider jobs")
-        self._write_logs(self._private_path, logs)
-
-    def _get_file_path(self, base_path: str) -> str:
-        """Return the full path to the log file."""
-        return os.path.join(base_path, f"{self._job_id}{self.FILE_EXTENSION}")
-
-    def _read_logs(self, base_path: str) -> Optional[str]:
-        """Read logs from the given path."""
-        log_path = self._get_file_path(base_path)
-        try:
-            # listdir refreshes the COS volume.
-            # The log file is written in the Scheduler and read from the Gateway, and there is 15m TTL.
-            os.listdir(base_path)
-            with open(log_path, "r", encoding=self.ENCODING) as log_file:
-                content = log_file.read()
-                logger.info(
-                    "[logs-storage] job_id=%s user_id=%s | Log written %s",
-                    self._job_id,
-                    self._user_id,
-                    log_path,
-                )
-                return content
-        except FileNotFoundError:
-            logger.info(
-                "[logs-storage] job_id=%s user_id=%s | Log file not found at %s",
-                self._job_id,
-                self._user_id,
-                log_path,
-            )
-            return None
-        except (UnicodeDecodeError, IOError) as e:
-            logger.error(
-                "[logs-storage] job_id=%s user_id=%s | Failed to read log file %s: %s",
-                self._job_id,
-                self._user_id,
-                log_path,
-                str(e),
-            )
-            return None
-
-    def _write_logs(self, base_path: str, logs: str) -> None:
-        """Write logs to the given path."""
-        log_path = self._get_file_path(base_path)
-        try:
-            with open(log_path, "w+", encoding=self.ENCODING) as log_file:
-                log_file.write(logs)
-                log_file.flush()
-        except (UnicodeDecodeError, IOError) as e:
-            logger.error(
-                "[logs-storage] job_id=%s user_id=%s | Failed to write log file %s: %s",
-                self._job_id,
-                self._user_id,
-                log_path,
-                str(e),
-            )
+        """Persist private logs for the job (provider jobs only)."""
